COMMIT FINAL DEL PROYECTO

+++ Back/Domain/Entities/Descuento.cs
@@ -5,7 +5,7 @@ namespace Domain.Entities
     public  class Descuento
     {
         public int IdDescuento { get; set; }
-        public double Valor { get; set; }
+        public Decimal Valor { get; set; }
         public string Descripcion { get; set; }
         public DateTime FechaInicio { get; set; }
         public DateTime FechaFin {  get; set; }
